Add support for MFA tuples in on_output config
This patch adds support for specifying an MFA tuple instead of a fun
when configuring `on_output`. This is very helpful when storing the
PropEr configuration in a config file, as many tools (such as rebar) use
`file:consult/1` to read configuration, making it impossible to use a
fun when specifying this setting.