Avoid useless shrinking when exceptions occur
Function same_fail_reason/2 was written in a way that two exceptions
were equal only when their exception reasons were exactly identical.
This had as a side effect that when exceptions included parts of the
input in them, shrinking was completely ineffective as it was trying
to generate the exact same exception reason. Change the code to only
check the tag of the exception reason.

NOTE: The code now assumes that exception reasons are either atoms
or tagged tuples. Perhaps this should be revised.