don't clean the stacktrace if the error is in the use of proper functions
Otherwise, the next error (not tested, is a simplification of an actually
failing test), with a wrong call to the invalid proper_fs:run_commands/1:

prop_bad_fsm() ->
    ?FORALL(
       Cmds, proper_fsm:commands(?MODULE),
       {H, S, R} = proper_fsm:run_commands(Cmds),
       ?WHENFAIL(report_error(H, S, R), R =:= ok)).

Will fail with undef and an empty stracktrace, which is confusing enough to
drain some time debugging it:

!
Failed: After 1 test(s).
An exception was raised: error:undef.
Stacktrace: [].
[]