Prevent crash when encountering the 'any' type argument
This fixes #94 where proper_typeserver would crash when resolving the
'erlang:timestamp()' type in OTP 17.x.

The underlying problem is that we were encountering the 'map()' type in
the erlang module with an argument of 'any'. At this point, the
typeserver only knew how to handle this situation for tuples.

Support has been extended to properly handle this argument for other
types, and the 'map()' type will now correctly be identified as
unsupported rather than causing a crash.

For example:

1> proper_typeserver:demo_translate_type(erlang, "map()").
{error,{unsupported_type,{type,1,map,any}}}