Modernize the callbacks of proper_fsm and proper_statem
The code, written long ago, used behaviour_info/1 to describe the
callbacks that the behaviour modules should export. Modernize the
code to use -callback attributes instead and describe the types of
the callbacks.

While at it, the types symb_var() and symb_call() were renamed to
symbolic_var() and symbolic_call() for consistency, and small errors
in the Edoc documentation were corrected.