Change put and erase to always return ok.
This makes it a bit more challenging for PropEr to find bugs
since it has less information to go on, so it makes for a
slightly more interesting example.

The change made it necessary to change the precondition for
get/1 to always return true since we need to be able to
get keys that should be undefined to detect if the erase
function is no longer working.