Support a map generator
Simply parametrize it from using two generators for keys and values.
The map is built from using a list of key/value pairs and converting
them.

This implementation was chosen for its simplicity, but also because it
allows to use a PropEr version that supports maps in an older Erlang
version without compile errors (nor runtime errors as long as the
generator is not used).