Convert targeted to proper interface (#201)
This pull request was done with 2 problems in mind:

1. Working to implement the Stateful TPBT of PropEr, I came across a problem with `?FORALL_TARGETED`, which turned out to come from not accepting `?TRAPEXIT` as a valid test. The problem was caused by [this line of code](https://github.com/proper-testing/proper/blob/3d74fde55450609bf11ba37bb28835375a2a5708/include/proper_common.hrl#L38), which does not accept any "non-immediate" boolean properties.
2. When calling the `?FORALL_TARGETED` macro, the messages after passing or failing are not in line with the rest of the PropEr interface, that is what was printed was always that the tests passed or failed "After 1 test(s).".

The changes made to the code, do not have any big implications on the `?FORALL_TARGETED` functionality, rather aim to fix the problems stated above. That is, `?FORALL_TARGETED` now is executing the number of tests defined by `numtests` and accepts "non-immediate" boolean properties.