Remove support for SFMT (#207)
Once upon a time, PropEr was enhanced with a configure option to employ
sfmt-erlang as an alternative to the random generator module, which was
(back then) quite slow(er).  Since then, Erlang/OTP has got support for
RNG in the form of the new `rand` module, which incorporates most (all?)
advantages that sfmt provides.

To simplify the code, support for sfmt-erlang is thereby discontinued.

It's not clear how used this configure option was in practice, anyway.