Build PropEr using rebar3 instead of own rebar
This set of changes allows PropEr to be built using rebar3 instead of
its own (patched) rebar version.  Besides being more modern, this
allows for easier integration of PropEr in code bases that use it as
a dependency.

Note that PropEr's version of rebar is still used for running the unit
tests, mainly due to being able to shut off more warnings than when
using rebar3.  However, this only affects PropEr developers, not users.

Addresses #204 and a request by propcheck.

Obsoletes #44 and most of #168.