Improve node handling
As previously stated, for stateful properties each worker
would need a node of his own to avoid clashing with
the other workers, whereas on stateless properties all the
workers could be performing tests on the same node.

This commit fixes last one by starting only one node if it
detects a stateless property is being tested, or a new node
per worker in the case of a stateful property.