Add some improvements to several functions
- `property_type/0` has a cleaner pattern matching when
identifying the kind a property belongs to.
- Sometimes when running PropEr's tests the master node would
count a message from a past property as a success/fail of the
current one being tested, so now there's an identifier (a timestamp)
for each property being tested.
- If a node was already up and running with a name that PropEr
uses with its nodes it would be stopped at the end of the testing
when `stop_nodes/0` was called, so now it differentiates if a node
was started or already running (and ensures the code is loaded in both cases).