Check values supplied as user options for validity
To avoid situations where `proper:quickcheck` crashes due to invalid
values supplied as user options (e.g., `{constraint_tries,-17}`,
`{numtests,0}`, etc.) the options are validated.  If they are invalid,
a new error / exception `{erroneous_option,Option}` is thrown.

While at it, slightly updated the documentation, reorganized options to
be listed alphabetically with each category, and also fixed a test which
was setting an option to an invalid value (to check a failure of a kind).