Add an Erlang abstract code generator
The Erlang/OTP team at Ericsson has used the abstract code generator
to find mostly trivial bugs in the Compiler, the Erlang meta
interpreter, the Erlang pretty printer, and more. It is provided as is
in the hope that it can be useful and also improved upon.

Many option names are shamelessly borrowed from the eqc_erlang_program
module of QuviQ's QuickCheck application.