Add an Erlang abstract code generator (#255)
The Erlang/OTP team at Ericsson has used this abstract code generator
to find bugs in the Compiler, the Erlang meta interpreter, the Erlang pretty
printer, and more.  It is provided as is in the hope that it can be useful to
others and also improved upon.

Many option names were shamelessly borrowed from  QuviQ's QuickCheck
eqc_erlang_program module.