Add `parallel-test-examples` target to Makefile and clean test suite
Instead of adding a workflow to run the test suite with parallel PropEr,
we should have a target in the Makefile to run the examples with it.

So, this commit removes all references to the `num_workers` option in
the suite except for the test case that runs the examples,
`examples_are_ok_test_/0`, as we do want to test them in parallel
and the test case will be removed in the future.