Move the testing of examples in the `examples` directories (#264)
Before GitHub actions, we separated some lengthy (and flaky) examples tests out of `proper_tests` and placed them in a separate test target in the `Makefile`.  For increased coverage, some of the examples were still tested in `proper_tests` as whole modules.

Both for consistency and in preparation for using the examples as tests for #263, we now have a _single_ place (GitHub action) where the complete set of examples is tested.