Extend PropEr with parallel execution (#263)
* Add a first approach to parellizing PropEr using processes

Modify the `user_opts` type and `opts` record to
allow the possibility of spawning multiple
processes to perform the tests

Update `inner_test` function: each process is
assigned a cerain number of tests to run from
the total count. The main process aggregates the
output received in case of success, otherwise
returns early with the error or fail

Some type specs were updated to reflect the changes
needed to use the `spawn_link_migrate` function

* Fix the case where there are more processes than tests

* Rename options and variables to be more clear

* Add the first steps towards a more safe parallelization

* Make PropEr use a node when parallelizing

This way, when using workers to distribute tests
among processes, PropEr will first create a node
for them to be spawned on, avoiding crashing the
whole VM if something goes wrong.

* Properly handle the processes termination

* Clean up the code a bit and fix dialyzer warnings

* Fix a pattern matching of received test results

* Avoid unneeded pattern matching and an extra line by using lists:foreach instead of lists:map

* Create as many nodes as processes will be spawned to avoid test collisions

When testing stateful programs, the node where
the tests were being performed would crash because
normally stateful programs use gen_servers or unique
elements, which made processes modify each other's
states.

To avoid that now each process is spawned on a
different node, which makes using many processes
a bit more expensive.

* Improve node handling

As previously stated, for stateful properties each worker
would need a node of his own to avoid clashing with
the other workers, whereas on stateless properties all the
workers could be performing tests on the same node.

This commit fixes last one by starting only one node if it
detects a stateless property is being tested, or a new node
per worker in the case of a stateful property.

* Fix `lists:zip/2` function_clause error

As `lists:zip/2` expects two lists of the same length,
it would crash on stateful properties when `numtests` was
less than `num_processes`. In that case, we only need to
spawn `numtests` processes (since otherwise we would end up
having idle workers).

* Fix wrong number of tests on property fail

Also did some refactor, mainly changin from
*processes* to *workers* (and so `num_processes`
to `num_workers`, etc), and added some documentation.

* Fix error handling on `perform` not being sent back

Also enforce a more strict pattern matching when trying
to determine if a property is stateless or stateful.

* Add more error handling

* Fix code not always being properly loaded on node

As PropEr internally works with anonymous functions to test
whether a property is satisfied or not, sometimes spawning them
on nodes would not work because not all the needed modules were loaded.

* Add some improvements to several functions

- `property_type/0` has a cleaner pattern matching when
identifying the kind a property belongs to.
- Sometimes when running PropEr's tests the master node would
count a message from a past property as a success/fail of the
current one being tested, so now there's an identifier (a timestamp)
for each property being tested.
- If a node was already up and running with a name that PropEr
uses with its nodes it would be stopped at the end of the testing
when `stop_nodes/0` was called, so now it differentiates if a node
was started or already running (and ensures the code is loaded in both cases).

* Force `?_assertBecomesN` to test in sequential PropEr

Will make those tests fail if not tested in a branch with
the `num_workers` option, though (currently `develop`).

* Force tests using `non_deterministic/1` to be run in sequential PropEr

* Make `?_assertBecomesN` to be run in sequential PropEr by default

And accordingly removing said option from its tests.

* Fix code loading and some tests not working when ran with workers

* Start a Cover server on each node if COVER env is found

In order to get the coverage from PropEr own tests while using
workers, Cover needs to know of the existing nodes to call
`cover:start/1` and cover:stop/1`.

* Fix code loading and statistic samples not being recollected

And while we are at it, change some wording too.

* Add size_at_nth_test/2 and perform/4

* Use a balanced test sequence distribution

This commit changes how the number of tests are passed to the workers
to a list of of test numbers they are starting on, e.g.:
A property with {numtests = 100, num_workers = 2} will result in
the following lists:

  - worker1: [0, 2, 4, ..., 98]
  - worker2: [1, 3, 5, ..., 99]

This way the workload is more balanced among workers, which avoids
having the biggest test inputs happen only on the last worker(s).

* Add optional strategy fun to override the default one

Also, get rid of old terminology in more places.

* Fix proper_tests to test it with 2 workers

* Update exists_tests and example_tests to pass with 2 workers

* Refactor logic around stateless/stateful properties to pure/impure

After a talk with Kostis, we agreed that instead of relying in pattern matching
against the result of the transformations of the forall property to determine
whether it was stateless or stateful, it would be a better idea to leave that
decision to the user and have them specify whether a property is:
  - `pure`, or side-effect free, that is; these properties will have all the
    workers running the tests in parallel in the same node
  - `impure`, not side-effect free; these properties will still have all the
    workers running the tests in paralle, but each in their own node to avoid
    them clashing with each other.

In addition, the docs have been updated to include all the new options, which are:
  - `num_workers`, sets the number of workers to be spawned to parallelize the tests
  - `strategy_fun`, overrides the default function that splits the tests among the workers
  - `pure` | `impure`, sets the property to that type

* Stop using nodes with stateless properties

* Avoid stopping nodes when `proper:module` was the entry point

* Add `parallel-test-examples` target to Makefile and clean test suite

Instead of adding a workflow to run the test suite with parallel PropEr,
we should have a target in the Makefile to run the examples with it.

So, this commit removes all references to the `num_workers` option in
the suite except for the test case that runs the examples,
`examples_are_ok_test_/0`, as we do want to test them in parallel
and the test case will be removed in the future.

* Rename make target to `test-parallel` and address review comments

* Disable `unknown` warning in Dialyzer

This commit can and will be dropped at a future point in time, and is
only being added to help continue testing parallel PropEr in Github Actions.

* Rename `num_workers` to `numworkers` to be consistent with the existing API

* Address review comments

* Refactor common code in parallel_perform/2 into a new function

Co-authored-by: pablocostass <pablo.costass@udc.es>