Suppress warnings on unused types
PropEr has the ability to turn type declarations into generators, and
thus allows types to appear in places where function calls would be
expected.  Alas, the Erlang compiler is unaware of this, and warns that
types are unused, as in the module below.  A simple fix, at least for
the time being, is to explicitly enable the compiler option that shuts
off compiler warnings about unused types.  Of course, a better option
would be for the Erlang compiler to be PropEr-aware...

-module(tt).
-include_lib("proper/include/proper.hrl").

-type t() :: 1..42.

prop_foo() ->
  ?FORALL(X, t(), X > 0).