Address deprecation warnings in OTP25
This commit addresses the warnings being emitted when compiling proper
in OTP25 related to the deprecation of the old `slave` module. To do so,
two new macros `?START_PEER_NODE/1` and `?STOP_PEER_NODE/1` have been
added, which handle the node setup and teardown properly depending on
the OTP version PropEr was compiled with.

I also took the chance to improve the naming in the codebase around
said nodes, and to make the code a bid cleaner.