Move cleanup logic into Makefile
PR #308 brought up the issue that the `clean` command of rebar3 had a
Unix-based script as a `post_hook` that could not be run on Windows.
Rather that adopting #308 and run the risk of leaving out some OSs out,
move all cleanup logic, including calling cleanup scripts, to the
Makefile instead. This should avoid `rebar3 clean` crashing on Windows.

While at it, move the `rebar3 clean` call from the `distclean` to the
`clean` target of Makefile, which seems more standard (and kosher).

Closes #308